Remove `_ProphetAdapter` base class and move logic directly to the `Prophet` forecaster implementation
